# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
  /login:
    x-swagger-router-controller: login
    post:
      parameters:
        - in: formData
          name: login
          type: string
          required: true
          description: User login
        - in: formData
          name: password
          type: string
          required: true
          description: User password
      produces:
        - application/json
      operationId: authentication
      responses:
        '200':
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    x-swagger-router-controller: users
    post:
      parameters:
        - in: formData
          name: login
          type: string
          required: true
          description: User login
        - in: formData
          name: password
          type: string
          required: true
          description: User password
        - in: formData
          name: firstName
          type: string
          required: true
          description: User first name
        - in: formData
          name: lastName
          type: string
          required: true
          description: User last name
      produces: 
        - application/json
      description: Create a new user
      operationId: createUser
      responses:
        '200':
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/id/{userId}:
    x-swagger-router-controller: users
    get:
      parameters:
        - in: path
          name: userId
          type: string
          required: true
          description: String ID of the user to get
      produces:
        - application/json
      description: Access to users
      operationId: getUserById
      responses:
        '200':
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      parameters:
        - in: path
          name: userId
          type: string
          required: true
          description: String ID of the user to put
        - in: formData
          name: firstName
          type: string
          required: true
        - in: formData
          name: lastName
          type: string
          required: true
      operationId: updateUserById
      responses:
        200:
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      parameters:
        - in: path
          name: userId
          type: string
          required: true
          description: String ID for delete user
      operationId: deleteUserById
      summary: Delete user
      description: Delete user by ID
      responses:
        200:
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
  /users/login/{userLogin}:
    x-swagger-router-controller: users
    get:
      parameters:
        - in: path
          name: userLogin
          type: string
          required: true
          description: Login of the user to get
      produces:
        - application/json
      description: Access to users
      operationId: getUserByLogin
      responses:
        '200':
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      parameters:
        - in: path
          name: userLogin
          type: string
          required: true
          description: String login of the user to put
        - in: formData
          name: firstName
          type: string
          required: true
        - in: formData
          name: lastName
          type: string
          required: true
      operationId: updateUserByLogin
      responses:
        200:
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      parameters:
        - in: path
          name: userLogin
          type: string
          required: true
          description: String ID for delete user
      operationId: deleteUserByLogin
      summary: Delete user
      description: Delete user by Login
      responses:
        200:
          description: OK
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string